added global menu class
