+++ app/src/main/java/com/example/smartshopper/LoginActivity.java
@@ -2,27 +2,14 @@ package com.example.smartshopper;
 import android.annotation.SuppressLint;
 import android.content.Intent;
 import android.os.Bundle;
-import android.util.Log;
-import android.view.MenuItem;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.ActionBarDrawerToggle;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.drawerlayout.widget.DrawerLayout;
 import com.example.smartshopper.common.PlatformHelpers;
 import com.example.smartshopper.utilities.LocalStorage;
-import com.example.smartshopper.utilities.NavigationDrawer;
-import com.google.android.material.internal.NavigationMenuItemView;
-import com.google.android.material.navigation.NavigationView;
-import java.util.Objects;
 
-public class LoginActivity extends AppCompatActivity {
-  DrawerLayout drawerLayout;
-  ActionBarDrawerToggle actionBarDrawerToggle;
-  NavigationDrawer navigationDrawer;
+public class LoginActivity extends MenuActivity {
   Button loginButton;
   PlatformHelpers platformHelpers;
   EditText emailAddressET;
@@ -33,14 +20,7 @@ public class LoginActivity extends AppCompatActivity {
   protected void onCreate(Bundle savedInstanceState) {
     super.onCreate(savedInstanceState);
     setContentView(R.layout.activity_login);
-    // https://www.geeksforgeeks.org/navigation-drawer-in-android/
-    // drawer layout instance to toggle the menu icon to open drawer and back button to close drawer
-    drawerLayout = findViewById(R.id.home_page_drawer_layout);
-    actionBarDrawerToggle = new ActionBarDrawerToggle(this, drawerLayout, R.string.nav_open, R.string.nav_close);
-    navigationDrawer = new NavigationDrawer(this, drawerLayout, actionBarDrawerToggle);
-    // to make the Navigation drawer icon always appear on the action bar
-    Objects.requireNonNull(getSupportActionBar()).setDisplayHomeAsUpEnabled(true);
-    setNavigationViewListener();
+
     platformHelpers = new PlatformHelpers(getApplicationContext());
     loginButton = findViewById(R.id.loginButton);
     loginButton.setOnClickListener(view -> {
@@ -58,43 +38,14 @@ public class LoginActivity extends AppCompatActivity {
 
     platformHelpers.checkEmail(emailAddress, password, userInterface -> {
       if (userInterface != null) {
-        try {
-          localStorage.setUser(userInterface);
-          NavigationMenuItemView signinButton = findViewById(R.id.nav_logout);
-          signinButton.setTitle("Sign Out");
-          Toast.makeText(getApplicationContext(), "Welcome back "+ userInterface.getUsername() + "!", Toast.LENGTH_LONG).show();
-          Intent mainActivityIntent = new Intent(this, MainActivity.class);
-          this.startActivity(mainActivityIntent);
-        } catch (Exception error) {
-          Log.v("error", error.toString());
-        }
-
+        localStorage.setUser(userInterface);
+        Toast.makeText(getApplicationContext(), "Welcome back "+ userInterface.getUsername() + "!", Toast.LENGTH_LONG).show();
+        Intent mainActivityIntent = new Intent(this, MainActivity.class);
+        this.startActivity(mainActivityIntent);
       } else {
         Toast.makeText(this, "Invalid Credentials", Toast.LENGTH_SHORT).show();
       }
     });
-  }
-  @Override
-  public boolean onOptionsItemSelected(@NonNull MenuItem item) {
-    if (actionBarDrawerToggle.onOptionsItemSelected(item)) {
-      return true;
-    }
-    return super.onOptionsItemSelected(item);
-  }
-
-  private void setNavigationViewListener() {
-    NavigationView navigationView = findViewById(R.id.navigation_view);
-    navigationView.setNavigationItemSelectedListener(navigationDrawer);
-  }
-
-  public void sendToLoginActivity(MenuItem item) {
-    if (localStorage.getCurrentUser() != null) {
-      Intent loginIntent = new Intent(this, LoginActivity.class);
-      startActivity(loginIntent);
-    }
-    else {
-      localStorage.signOut();
-    }
   }
 
   public void sendToForgotPasswordActivity(View view) {
