+++ app/src/main/res/menu/navigation_menu.xml
@@ -1,5 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
+<menu
+    android:id="@+id/nav_menu"
+    xmlns:android="http://schemas.android.com/apk/res/android">
 
     <item
         android:id="@+id/nav_account"
